// src/app/interfaces/verdict.ts
export interface IVerdict {
  key?: string;
  status?: any;
  page_key?: number;
  create_date?: Date;
  create_by?: any;
  update_date?: Date;
  update_by?: any;
  crime?: any;
  verdict_no: string;
  judge?: any;
  judge_ask_date?: any;
  judge_ask_date_key?: number;
  judgment_writer?: any;
  judgment_writer_date?: any;
  judgment_writer_date_key?: number;
  judgment_status?: any;
}